Warnings are Errors (#437)
This PR fixes all current warnings in fVDB. The goal is that by default, the build should only print messages when there is something wrong. It is quiet by default and there are no warnings.

 -  Fixes existing warnings that occur without -Wall
 - Fixes additional (many more) warnings that occur with -Wall. Mostly signed/unsigned comparisons and unused variables.
 - Enables -Wall -Werror for CXX options
 - Enables -Werror=all-warnings -XCompiler=-Wall,-Werror for CUDA options.
 - Disables a bunch of #pragma message output by default
 - Makes the build quiet by default rather than verbose.

Once this is merged, compilation will fail if there are any warnings.

To do a verbose build now, add -v for verbose pip output, and -C build.verbose=true for verbose CMake output (print every compilation command line).

---------

+++ fvdb/src/detail/ops/ActiveVoxelsInBoundsMask.cu
@@ -169,8 +169,8 @@ ActiveVoxelsInBoundsMask(const GridBatchImpl &batchHdl,
         torch::empty({batchHdl.batchSize(), 2, 3},
                      torch::TensorOptions().dtype(torch::kInt32).device(batchHdl.device()));
 
-    for (size_t batchIdx = 0; batchIdx < batchHdl.batchSize(); batchIdx++) {
-        for (size_t dimIdx = 0; dimIdx < 3; dimIdx++) {
+    for (int64_t batchIdx = 0; batchIdx < batchHdl.batchSize(); batchIdx++) {
+        for (int32_t dimIdx = 0; dimIdx < 3; dimIdx++) {
             batchBboxes[batchIdx][0][dimIdx] = bboxMins[batchIdx][dimIdx];
             batchBboxes[batchIdx][1][dimIdx] = bboxMaxs[batchIdx][dimIdx];
         }
