+++ fvdb/src/detail/ops/ReadFromDense.cu
@@ -58,7 +58,7 @@ readFromDenseCPU(const GridBatchImpl::Accessor<GridType> &gridHandle,
                  torch::TensorAccessor<ScalarType, 2> outSparseTensor,
                  bool ignoreMasked,
                  bool isContiguous) {
-    for (size_t bi = 0; bi < gridHandle.batchSize(); bi += 1) {
+    for (int64_t bi = 0; bi < gridHandle.batchSize(); bi += 1) {
         const nanovdb::NanoGrid<GridType> *grid = gridHandle.grid(bi);
         const nanovdb::Coord denseOrigin(
             denseOrigins[bi][0], denseOrigins[bi][1], denseOrigins[bi][2]);
